README.txt added, describing the installation and set-up processes. Also describing to use the FAQ or forum when you have problems and finally to read minsoc.pdf (the documentation) to understand the system after everything works, to know what to do next.
Clean-up of sw/utils, removing unused sources and files. 

Update of sw/drivers/eth.c, direct casting to avoid compile warnings. 

FAQ extended and with more links to the threads giving the solutions. 

Synthesis examples inform that the firmware tweak for Spartan3E Starter Kit using Ethernet only works for system version up to 35. 

minsoc.pdf and webpage now uses FPGA generic and specific codes instead of FPGA independent and dependent codes, because it suits it better.

+++ sw/drivers/eth.c
@@ -34,11 +34,11 @@ void eth_init()
 									 | (OWN_MAC_ADDRESS_1 << 8) | OWN_MAC_ADDRESS_0;		//mac address rest
 
 	//configure TXBD0
-	REG32(ETH_BASE + ETH_TXBD0H) = eth_tx_packet;		//address used for tx_data
+	REG32(ETH_BASE + ETH_TXBD0H) = (unsigned long)eth_tx_packet;		//address used for tx_data
 	REG32(ETH_BASE + ETH_TXBD0L) = TX_READY;	//length = 0 | PAD & CRC = 1 | IRQ & WR = 1
 
 	//configure RXBD0
-	REG32(ETH_BASE + ETH_RXBD0H) = eth_rx_packet;		//address used for tx_data
+	REG32(ETH_BASE + ETH_RXBD0H) = (unsigned long)eth_rx_packet;		//address used for tx_data
 	REG32(ETH_BASE + ETH_RXBD0L) = RX_READY;	//len = 0 | IRQ & WR = 1 | EMPTY = 1
 
 	//set txdata
